Migrate Query.schemaCoordinateStats
